Feat: 공용 헤더 사용 및 로직 변경
